# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: longue-vue CMDB API
  version: 0.1.0
  summary: Kubernetes-centric CMDB aligned with the ANSSI SecNumCloud framework.
  description: |
    REST API for **longue-vue**, a CMDB for Kubernetes environments aligned with the
    ANSSI SecNumCloud (SNC) qualification framework.

    This version (0.1.0) is a walking skeleton covering:
      - Liveness and readiness probes
      - `Cluster` resource CRUD

    Subsequent versions will add Node, Namespace, Workload, Pod, and other
    Kubernetes entity kinds mapped onto the ANSSI cartography layers.

    Errors follow RFC 7807 (`application/problem+json`).

    Interactive API documentation is available at `/docs/` for authenticated
    operators; sign in via `/ui/login` first to use "Try it out".
servers:
  - url: /
    description: Current origin (recommended for browser-based "Try it out" via /docs/)
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/eol/extract:
    get:
      tags: [eol]
      summary: Download EOL Dashboard rows as CSV or JSON
      description: |
        Flattens longue-vue.io/eol.* annotations from clusters, nodes
        and virtual machines into one row per (entity, product). Optional
        `entity_type` and `status` filters mirror the dashboard's chips.
      operationId: eolExtract
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - in: query
          name: format
          required: true
          schema:
            type: string
            enum: [csv, json]
        - in: query
          name: entity_type
          schema:
            type: string
            enum: [cluster, node, vm, workload]
        - in: query
          name: status
          schema:
            type: string
            enum: [eol, approaching_eol, supported, unknown]
      responses:
        '200':
          description: Extract body.
          headers:
            X-Longue-Vue-Truncated:
              schema:
                type: string
                enum: ['true']
            Content-Disposition:
              schema: { type: string }
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EolExtractRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/search/extract:
    get:
      tags: [search]
      summary: Download Search results as CSV or JSON
      description: |
        Returns every Workload / Pod / Virtual Machine matching the
        cross-entity search query `q`, restricted to the entity kind
        named in `kind`. The full filtered result set is streamed, capped
        at `LONGUE_VUE_EXTRACT_MAX_ROWS` (default 50 000). When the cap
        is hit the response carries `X-Longue-Vue-Truncated: true` and
        the file is still well-formed CSV / JSON.
      operationId: searchExtract
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
            maxLength: 256
        - in: query
          name: kind
          required: true
          schema:
            type: string
            enum: [workloads, pods, virtual_machines]
        - in: query
          name: format
          required: true
          schema:
            type: string
            enum: [csv, json]
        - in: query
          name: application
          required: false
          description: |
            ADR-0029 §2.4. Narrows the result set to entities whose linked
            Application's name matches the substring (case-insensitive).
            Affects workloads + VMs directly; pods inherit via their
            owning workload's link. AND-combined with `q`.
          schema:
            type: string
            maxLength: 256
      responses:
        '200':
          description: Extract body. CSV is UTF-8 with a BOM and CRLF endings; JSON is a flat array.
          headers:
            X-Longue-Vue-Truncated:
              schema:
                type: string
                enum: ['true']
              description: Present and "true" when the row cap was hit.
            Content-Disposition:
              schema: { type: string }
              description: 'attachment; filename="longue-vue-search-<kind>-<q-slug>-<ts>.<csv|json>"'
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                anyOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/WorkloadExtractRow'
                  - type: array
                    items:
                      $ref: '#/components/schemas/PodExtractRow'
                  - type: array
                    items:
                      $ref: '#/components/schemas/VirtualMachineExtractRow'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /v1/clusters/{id}/flow-matrix:
    get:
      operationId: getClusterFlowMatrix
      summary: Synthesized reference-vs-discovered flow matrix for a cluster
      description: >-
        Read-time synthesis (ADR-0036) comparing the cluster's discovered
        perimeter security-group rules and internal NetworkPolicy rules against
        the operator-declared reference matrix. Gated by the admin-settings
        flow_matrix_enabled toggle; returns 409 when disabled.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterFlowMatrix"
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
  /v1/clusters/{id}/flow-matrix/extract:
    get:
      tags: [flow-matrix]
      operationId: clusterFlowMatrixExtract
      summary: Download a cluster's flow-matrix synthesis as CSV or JSON
      description: |
        Flattens the read-time perimeter + internal flow-matrix synthesis
        (ADR-0037) into one row per synthesized flow. Read scope, gated by
        the admin-settings flow_matrix_enabled toggle (409 when off), capped
        at `LONGUE_VUE_EXTRACT_MAX_ROWS` (default 50 000). When the cap is
        hit the response carries `X-Longue-Vue-Truncated: true`. Audited
        (SecNumCloud ch.8 evidence).
      security:
        - BearerAuth: [read]
        - SessionCookie: [read]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - in: query
          name: format
          required: false
          schema:
            type: string
            enum: [csv, json]
            default: csv
      responses:
        '200':
          description: Extract body. CSV is UTF-8 with a BOM and CRLF endings; JSON is an envelope with an items array.
          headers:
            X-Longue-Vue-Truncated:
              schema:
                type: string
                enum: ['true']
              description: Present and "true" when the row cap was hit.
            Content-Disposition:
              schema: { type: string }
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: token
      description: |
        API token issued out-of-band. Send as `Authorization: Bearer <token>`.

        Scopes granted to a token:
          - `read`   — list and get operations
          - `write`  — create and update (implies `read` at the caller's discretion)
          - `delete` — remove resources
          - `admin`  — all of the above; reserved for future admin-only endpoints

        Tokens carrying the `admin` scope implicitly satisfy any scope required
        by an operation.

        Per ADR-0007, machine tokens are created by an admin in the UI
        (see `POST /v1/admin/tokens`) rather than injected through env vars.

    SessionCookie:
      type: apiKey
      in: cookie
      name: longue_vue_session
      description: |
        Human session cookie set by `POST /v1/auth/login`. HttpOnly,
        SameSite=Strict, 8-hour sliding expiry. Per ADR-0007.

        Every operation accepts either `BearerAuth` or `SessionCookie` —
        scope enforcement reads from the caller's role (for sessions) or
        token scopes (for bearer), so the two paths share the same
        downstream authorisation check.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Token is valid but lacks the scope required by the operation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Conflict:
      description: Conflict with current state (e.g., duplicate name).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      description: RFC 7807 problem details.
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri
        errors:
          type: array
          description: Field-level validation errors, when applicable.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    FlowMatrixPortRange:
      type: object
      description: Inclusive port span. from omitted means "any port".
      required: [protocol]
      properties:
        protocol: { type: string, enum: [tcp, udp, icmp, any] }
        from:     { type: integer }
        to:       { type: integer }
    FlowMatrixEndpoint:
      type: object
      required: [kind, name]
      properties:
        kind: { type: string, enum: [workload, service, namespace, endpoint_group, cidr] }
        name: { type: string }
        cidr: { type: string }
    FlowMatrixSource:
      type: object
      description: One contributing rule for a synthesized flow.
      required: [kind, id, summary]
      properties:
        kind:    { type: string, enum: [sg_rule, netpol_rule, reference] }
        id:      { type: string }
        summary: { type: string }
    FlowMatrixFlow:
      type: object
      description: One synthesized cell of the flow matrix.
      required: [layer, direction, src, dst, ports, state, sources]
      properties:
        layer:        { type: string, enum: [perimeter, internal] }
        direction:    { type: string, enum: [ingress, egress] }
        src:          { $ref: "#/components/schemas/FlowMatrixEndpoint" }
        dst:          { $ref: "#/components/schemas/FlowMatrixEndpoint" }
        ports:        { type: array, items: { $ref: "#/components/schemas/FlowMatrixPortRange" } }
        state:        { type: string, enum: [conforme, non_declare, manquant, large_ouvert] }
        reference_id: { type: string }
        sources:      { type: array, items: { $ref: "#/components/schemas/FlowMatrixSource" } }
    FlowMatrixWarning:
      type: object
      description: A non-fatal synthesis issue (e.g. a dangling reference).
      required: [kind]
      properties:
        kind:         { type: string }
        reference_id: { type: string }
        ref:          { type: string }
    ClusterFlowMatrix:
      type: object
      description: Full per-cluster flow-matrix synthesis result (ADR-0036).
      required: [cluster_id, perimeter, internal, warnings]
      properties:
        cluster_id: { type: string }
        perimeter:  { type: array, items: { $ref: "#/components/schemas/FlowMatrixFlow" } }
        internal:   { type: array, items: { $ref: "#/components/schemas/FlowMatrixFlow" } }
        warnings:   { type: array, items: { $ref: "#/components/schemas/FlowMatrixWarning" } }
...
    EolExtractRow:
      type: object
      required: [entity_type, entity_id, entity_name, product, cycle, status]
      properties:
        entity_type: { type: string, enum: [cluster, node, vm, workload] }
        entity_id: { type: string }
        entity_name: { type: string }
        cluster: { type: string }
        product: { type: string }
        cycle: { type: string }
        status: { type: string, enum: [eol, approaching_eol, supported, unknown] }
        eol_date: { type: string }
        latest: { type: string }
        latest_available: { type: string }
        support: { type: string }
        checked_at: { type: string }
    PodExtractRow:
      type: object
      required: [cluster, namespace, name]
      properties:
        id: { type: string, format: uuid }
        cluster: { type: string }
        namespace: { type: string }
        name: { type: string }
        workload_kind: { type: string }
        workload_name: { type: string }
        image_matches: { type: string }
        phase: { type: string }
        node: { type: string }
        updated_at: { type: string, format: date-time }
    VirtualMachineExtractRow:
      type: object
      required: [name]
      properties:
        id: { type: string, format: uuid }
        cloud_account: { type: string }
        region: { type: string }
        name: { type: string }
        display_name: { type: string }
        role: { type: string }
        power_state: { type: string }
        image_id: { type: string }
        image_name: { type: string }
        applications_matched: { type: string }
        updated_at: { type: string, format: date-time }
    WorkloadExtractRow:
      type: object
      required: [cluster, namespace, kind, name]
      properties:
        id: { type: string, format: uuid }
        cluster: { type: string }
        namespace: { type: string }
        kind: { type: string }
        name: { type: string }
        image_matches: { type: string }
        replicas: { type: integer }
        ready_replicas: { type: integer }
        updated_at: { type: string, format: date-time }

    # ── ADR-0029: Application + ApplicationBlock ─────────────────────────────
...